updated PlaylistIdNotFoundException Class

+++ src/main/java/com/niit/jdp/repository/PlaylistRepository.java
@@ -6,6 +6,7 @@
  */
 package com.niit.jdp.repository;
 
+import com.niit.jdp.exeception.PlaylistIdNotFoundException;
 import com.niit.jdp.model.Song;
 
 import java.sql.Connection;
@@ -26,7 +27,7 @@ public class PlaylistRepository {
 	 * @param songNumbers  This is the string of song numbers that the user wants to add to the playlist.
 	 * @return A boolean value.
 	 */
-	public boolean addSongsInPlaylist(Connection connection, String playlistName, String songNumbers) throws SQLException {
+	public boolean createNewPlaylist(Connection connection, String playlistName, String songNumbers) throws SQLException {
 		// convert string to array
 		char[] songNumbersArray = songNumbers.toCharArray();
 
@@ -59,27 +60,31 @@ public class PlaylistRepository {
 	 * @param playlistId The id of the playlist you want to get the songs from.
 	 * @return A list of songs
 	 */
-	public List<Song> getAllSongsInPlaylist(Connection connection, int playlistId) throws SQLException {
-		// Create an object of SongRepository class
-		SongRepository songRepository = new SongRepository();
-		// Create a list of songs
-		List<Song> songList = new ArrayList<>();
+	public List<Song> getAllSongsInPlaylist(Connection connection, int playlistId) throws SQLException, PlaylistIdNotFoundException {
+		if (playlistId < 0) {
+			throw new PlaylistIdNotFoundException("Id Cannot be Negative");
+		} else {
+			// Create an object of SongRepository class
+			SongRepository songRepository = new SongRepository();
+			// Create a list of songs
+			List<Song> songList = new ArrayList<>();
 
-		// store the ids in a string
-		String idsFromPlaylist = getIdsFromPlaylist(connection, playlistId);
-		// convert the string to array seperated by (,)
-		String[] idInString = idsFromPlaylist.split(",");
+			// store the ids in a string
+			String idsFromPlaylist = getIdsFromPlaylist(connection, playlistId);
+			// convert the string to array seperated by (,)
+			String[] idInString = idsFromPlaylist.split(",");
 
-		// iterate over array to convert string array to integer array
-		int[] idInInteger = Arrays.stream(idInString).mapToInt(Integer::parseInt).toArray();
+			// iterate over array to convert string array to integer array
+			int[] idInInteger = Arrays.stream(idInString).mapToInt(Integer::parseInt).toArray();
 
-		// iterate over integer array and store the variable in list
-		for (int ids : idInInteger) {
-			Song song = songRepository.geyById(connection, ids);
-			songList.add(song);
+			// iterate over integer array and store the variable in list
+			for (int ids : idInInteger) {
+				Song song = songRepository.geyById(connection, ids);
+				songList.add(song);
+			}
+			// return the list of songs
+			return songList;
 		}
-		// return the list of songs
-		return songList;
 	}
 
 
@@ -118,18 +123,22 @@ public class PlaylistRepository {
 	 * @param playlistId The id of the playlist
 	 * @return A boolean value.
 	 */
-	public boolean deletePlaylist(Connection connection, int playlistId) throws SQLException {
-		// write the query to delete the playlist
-		String query = "DELETE FROM `jukebox`.`playlists` WHERE (`id` = ?);";
+	public boolean deletePlaylist(Connection connection, int playlistId) throws SQLException, PlaylistIdNotFoundException {
+		if (playlistId < 0) {
+			throw new PlaylistIdNotFoundException("Id Cannot be Negative");
+		} else {
+			// write the query to delete the playlist
+			String query = "DELETE FROM `jukebox`.`playlists` WHERE (`id` = ?);";
 
-		// create a prepared statement object
-		try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
+			// create a prepared statement object
+			try (PreparedStatement preparedStatement = connection.prepareStatement(query)) {
 
-			// set the values to the query
-			preparedStatement.setInt(1, playlistId);
+				// set the values to the query
+				preparedStatement.setInt(1, playlistId);
 
-			// execute the query
-			return preparedStatement.executeUpdate() > 0;
+				// execute the query
+				return preparedStatement.executeUpdate() > 0;
+			}
 		}
 	}
